Rename location/history_location to locations/history_locations
And location/none_location to locations/none
