+++ locations/history/link_delegate.js
