+++ test/test_app/js/test-app.js
@@ -28,7 +28,7 @@ define(function (require) {
   var Promise = require("cherrytree/vendor/promise");
   var Router = require("cherrytree");
   var Route = require("cherrytree/route");
-  var HistoryLocation = require("cherrytree/location/history_location");
+  var HistoryLocation = require("cherrytree/locations/history");
 
   var template = function (name) {
     return _.template("<div>" + $("script#" + name).html() + "</div>");
